pyBankRapport/requirements.txt
configobj>=5.0.6
python-dateutil>=2.8.2
timeloop>=1.0.2
pysftp>=0.2.9
paramiko>=2.9.2